Mapa en vista diferente que se inicializa con la ubicacion de Cuba
