+++ app/src/main/AndroidManifest.xml
@@ -12,10 +12,14 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Labiofamandroid"
         tools:targetApi="31">
+        <activity
+            android:name=".MapActivity"
+            android:exported="false" />
+
         <meta-data
             android:name="com.google.android.geo.API_KEY"
-            android:value="@string/google_map_api_key">
-        </meta-data>
+            android:value="@string/google_map_api_key"></meta-data>
+
         <activity
             android:name=".MainActivity"
             android:exported="true">
