+++ app/src/main/java/com/example/labiofam_android/MapActivity.kt
@@ -1,0 +1,26 @@
+package com.example.labiofam_android
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.OnMapReadyCallback
+import com.google.android.gms.maps.SupportMapFragment
+import com.google.android.gms.maps.model.LatLng
+
+class MapActivity : AppCompatActivity(), OnMapReadyCallback {
+    private var mGoogleMap: GoogleMap? = null
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_map)
+        val mapFragment = supportFragmentManager.findFragmentById(R.id.mapFragment)
+                as SupportMapFragment
+        mapFragment.getMapAsync(this)
+    }
+
+    override fun onMapReady(googleMap: GoogleMap) {
+        mGoogleMap = googleMap
+        val initialPosition = LatLng(21.5218, -77.7812)
+        mGoogleMap!!.moveCamera(CameraUpdateFactory.newLatLng(initialPosition))
+    }
+}
