+++ app/src/main/res/layout/activity_main.xml
@@ -6,15 +6,19 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/mapFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        class="com.google.android.gms.maps.SupportMapFragment"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-    </androidx.fragment.app.FragmentContainerView>
+
+    <Button
+        android:id="@+id/pv_btn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="168dp"
+        android:text="Puntos de Venta"
+        android:onClick="openPV"
+        android:textColor="@color/black"
+        style="?android:attr/borderlessButtonStyle"
+        app:layout_constraintEnd_toStartOf="@+id/appCompatTextView"
+        tools:layout_editor_absoluteY="16dp"
+        tools:ignore="MissingConstraints,UnknownId" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
+
